requirements.txt
facebook-business
gspread
google-auth
twilio
PyYAML
python-dotenv
anthropic
opencv-python>=4.11.0
